misc: move shared code to plugins/internal
